Raster IO (read)

Read a window of data from a GDAL raster source. The first argument is the source name and the second is a 6-element window of offset, source dimension, and output dimension.

raster_io(x, band = 1, window, resample = "nearestneighbour", ...,
  sds = NULL)

Arguments

x
data source

band
index of which band to read

window
src_offset, src_dim, out_dim

resample
resampling method used (see details)

...
reserved

sds
index of subdataset to read (usually 1)

Details

This is analgous to the rgdal function readGDAL with its arguments offset, region.dim and output.dim. There's no semantic wrapper for this in vapour, but see https://github.com/hypertidy/lazyraster for one approach.

Resampling options will depend on GDAL version, but currently 'NearestNeighbour' (default), 'Average', 'Bilinear', 'Cubic', 'CubicSpline', 'Gauss', 'Lanczos', 'Mode' are potentially available. These are compared internally by converting to lower-case. Detailed use of this is barely tried or tested with vapour, but is a standard facility used in GDAL. Easiest way to compare results is with gdal_translate.

There is no write support in vapour.

Examples

f <- system.file("extdata", "sst.tif", package = "vapour")
## a 5*5 window from a 10*10 region
raster_io(f, window = c(0, 0, 10, 10, 5, 5))

#>  [1] 285.743 285.786 285.999 286.064 286.038 285.763 285.698 285.831 286.265
#> [10] 286.134 286.070 285.931 286.092 286.450 286.501 286.264 286.076 285.915
#> [19] 286.335 286.702 285.870 285.940 286.252 286.549 287.142

raster_io(f, window = c(0, 0, 10, 10, 5, 5), resample = "Lanczos")

#>  [1] 285.7112 285.7088 285.9000 286.0153 285.9586 285.7220 285.7106 285.7733
#>  [9] 286.1091 286.2422 286.0036 285.8829 285.9075 286.3403 286.4126 286.2443
#> [17] 286.1346 285.9894 286.3176 286.6353 286.0227 285.9557 286.0615 286.3574
#> [25] 286.8593

## find the information first
ri <- raster_info(f)
str(matrix(raster_io(f, window = c(0, 0, ri$dimXY, ri$dimXY)), ri$dimXY[1]))

#>  num [1:143, 1:286] 286 286 286 286 286 ...

## the method can be used to up-sample as well
str(matrix(raster_io(f, window = c(0, 0, 10, 10, 15, 25)), 15))

#>  num [1:15, 1:25] 286 286 286 286 286 ...