vapour

A lightweight GDAL API package for R.

Details

Lightweight means we access parts of the GDAL API as near as possible to their native usage. GDAL is not a particularly lightweight library, but provide a very nice abstraction over format details for a very large number of different formats.

Functions for raster and vector sources are included.

vapour_raster_info	structural metadata of a source

vapour_read_raster	read data direct from a window of a raster band source

vapour_sds_names	list individual raster sources in a source containing subdatasets

vapour_raster_info

vapour_layer_names	list names of vector layers in a data source

vapour_read_names	read the 'names' of features in a layer, the 'FID'

vapour_read_attributes	read attributes of features in a layer, the columnar data associated with each geometry

vapour_read_extent	read the extent, or bounding box, of geometries in a layer

vapour_read_geometry	read geometry in binary (blob, WKB) form

vapour_read_geometry_text	read geometry in text form, various formats

vapour_layer_names

vapour_read_geometry_cpp is a general function that will return different types of output for different inputs and is used by vapour_read_extent, vapour_read_geometry and vapour_read_geometry_text.

As far as possible vapour aims to minimize the level of interpretation provided for the functions, so that developers can choose how things are implemented. Functions return raw lists or vectors rather than data frames or classed types.