vapour

A lightweight GDAL API package for R.

Details

Provides low-level access to 'GDAL' functionality for R packages. The aim is to minimize the level of interpretation put on the 'GDAL' facilities, to enable direct use of it for a variety of purposes. 'GDAL' is the 'Geospatial Data Abstraction Library' a translator for raster and vector geospatial data formats that presents a single raster abstract data model and single vector abstract data model to the calling application for all supported formats https://gdal.org/.

Lightweight means we access parts of the GDAL API as near as possible to their native usage. GDAL is not a lightweight library, but provide a very nice abstraction over format details for a very large number of different formats.

Functions for raster and vector sources are included.

#'

vapour_all_drivers	list of all available drivers, with type and features

vapour_gdal_version	report version of GDAL in use

vapour_raster_gcp	return internal ground control points, if present

vapour_raster_info	structural metadata of a source

vapour_read_raster	read data direct from a window of a raster band source

vapour_sds_names	list individual raster sources in a source containing subdatasets

vapour_driver	report name of the driver used for a given source

vapour_geom_summary	report simple properties of each feature geometry

vapour_layer_names	list names of vector layers in a data source

vapour_read_names	read the 'names' of features in a layer, the 'FID'

vapour_read_attributes	read attributes of features in a layer, the columnar data associated with each geometry

vapour_read_extent	read the extent, or bounding box, of geometries in a layer

vapour_read_geometry	read geometry in binary (blob, WKB) form

vapour_read_geometry_text	read geometry in text form, various formats

vapour_report_attributes	report internal type of each attribute by name

As far as possible vapour aims to minimize the level of interpretation provided for the functions, so that developers can choose how things are implemented. Functions return raw lists or vectors rather than data frames or classed types.