GDAL geometry extent

Read a GDAL geometry summary as just the native bounding box, the four numbers xmin, xmax, ymin, ymax in the usual simple convention.

vapour_read_extent(dsource, layer = 0L, sql = "")

Arguments

dsource
data source name (path to file, connection string, URL)

layer
integer of layer to work with, defaults to the first (0)

sql
if not empty this is executed against the data source (layer will be ignored)

Examples

file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
vapour_read_extent(mvfile)

#> [[1]]
#> [1]  448353.3  457705.8 5386605.8 5397351.8
#> 
#> [[2]]
#> [1]  453543.6  459318.0 5403972.4 5412505.1
#> 
#> [[3]]
#> [1]  454839.8  461042.0 5411562.5 5417891.6
#> 
#> [[4]]
#> [1]  461504.9  476213.0 5410911.3 5424854.2
#> 
#> [[5]]
#> [1]  471572.5  483157.2 5417109.6 5424645.0
#> 
#> [[6]]
#> [1]  491638.3  494048.2 5417261.7 5419330.7
#> 
#> [[7]]
#> [1]  508512.4  510057.3 5408708.8 5409648.5
#> 
#> [[8]]
#> [1]  435993.1  444516.8 5392899.6 5402213.0
#> 
#> [[9]]
#> [1]  453298.4  462671.3 5383592.2 5393869.2
#> 
#> [[10]]
#> [1]  457591.8  473356.6 5377059.2 5397051.3
#> 
#> [[11]]
#> [1]  465137.3  477417.7 5394644.9 5408531.0
#> 
#> [[12]]
#> [1]  495773.4  507915.4 5396426.7 5405789.0
#> 
#> [[13]]
#> [1]  489938.9  499711.7 5410676.5 5418328.0
#> 
#> [[14]]
#> [1]  459055.7  471904.7 5418849.2 5425154.3
#> 
#> [[15]]
#> [1]  455980.3  463978.7 5392869.8 5399258.9
#> 
#> [[16]]
#> [1]  475888.9  493826.3 5379076.9 5389494.4
#> 
#> [[17]]
#> [1]  480223.5  490958.3 5397051.9 5414285.8
#> 
#> [[18]]
#> [1]  485858.0  494029.1 5407194.0 5418266.9
#> 
#> [[19]]
#> [1]  505330.1  508582.3 5407832.4 5411813.2
#> 
#> [[20]]
#> [1]  412718.7  433693.8 5364101.9 5392659.1
#> 
#> [[21]]
#> [1]  418130  431437 5352613 5368181
#> 
#> [[22]]
#> [1]  452420.0  460772.2 5395775.1 5404446.8
#> 
#> [[23]]
#> [1]  459778.6  470225.2 5398390.6 5405474.0
#> 
#> [[24]]
#> [1]  471218.9  485529.5 5383849.5 5395087.3
#> 
#> [[25]]
#> [1]  457719.0  469430.2 5407328.1 5420616.2
#> 
#> [[26]]
#> [1]  486794.4  497211.2 5391895.7 5400794.2
#> 
#> [[27]]
#> [1]  488189.9  498091.5 5396973.3 5409378.0
#> 
#> [[28]]
#> [1]  479267.6  487580.6 5412326.3 5419667.5
#> 
#> [[29]]
#> [1]  487082.1  491199.8 5405846.1 5410168.0
#> 
#> [[30]]
#> [1]  507191.9  508681.9 5409571.8 5411550.2
#> 
#> [[31]]
#> [1]  432998.3  441412.9 5402890.6 5406594.9
#> 
#> [[32]]
#> [1]  457812.6  463400.1 5408117.3 5415279.1
#> 
#> [[33]]
#> [1]  470867.5  483434.3 5404320.4 5419938.1
#> 
#> [[34]]
#> [1]  481799.9  490524.1 5388711.3 5398922.0
#> 
#> [[35]]
#> [1]  493482.6  506412.5 5403348.9 5412391.3
#> 
#> [[36]]
#> [1]  416790.0  424280.7 5390142.2 5396737.7
#> 
#> [[37]]
#> [1]  432185.6  439290.6 5396540.6 5404520.9
#> 
#> [[38]]
#> [1]  462417.0  469411.5 5391796.8 5398603.8
#> 
#> [[39]]
#> [1]  473322.8  481799.9 5392389.8 5399468.9
#> 
#> [[40]]
#> [1]  466297.7  472555.8 5407297.4 5412734.5
#> 
#> [[41]]
#> [1]  488006.2  497238.9 5385525.8 5393899.4
#> 
#> [[42]]
#> [1]  500476.3  505427.0 5409887.5 5413864.0
#> 
#> [[43]]
#> [1]  422717.5  443693.4 5364752.4 5398968.7
#> 
#> [[44]]
#> [1]  436392.9  455344.9 5390196.5 5405568.8
#> 
#> [[45]]
#> [1]  448464.2  450389.7 5404211.8 5404796.7
#> 
#> [[46]]
#> [1]  459377.3  465846.6 5397255.3 5404092.7
#> 
#> [[47]]
#> [1]  475501.6  482058.2 5400269.4 5408109.1
#> 
#> [[48]]
#> [1]  476793.1  484115.0 5394878.4 5400904.9
#> 
#> [[49]]
#> [1]  495935.5  499732.6 5389588.9 5396934.4
#> 
#> [[50]]
#> [1]  425270.0  449448.4 5353987.4 5383046.1
#> 
#> [[51]]
#> [1]  439927.1  462874.4 5358222.8 5393865.2
#> 
#> [[52]]
#> [1]  468038.6  477379.7 5378281.3 5388344.2
#> 
#> [[53]]
#> [1]  457540.6  466788.4 5403285.3 5409103.7
#> 
#> [[54]]
#> [1]  495086.2  500371.5 5393786.4 5399069.3
#> 
#> [[55]]
#> [1]  503880.4  507861.2 5396801.3 5404336.2
#> 
#> [[56]]
#> [1]  502594.7  507481.2 5402751.6 5406381.9
#> 
#> [[57]]
#> [1]  506074.8  509426.1 5404304.6 5408099.0
#> 
#> [[58]]
#> [1]  506637.3  511687.5 5405460.0 5409603.3
#>