Read GDAL geometry as blob

Simple read of geometry-only as WKB format.

vapour_read_geometry(dsource, layer = 0L, sql = "")

Arguments

dsource
data source name (path to file, connection string, URL)

layer
integer of layer to work with, defaults to the first (0)

sql
if not empty this is executed against the data source (layer will be ignored)

Format

indicate text output format, available are "json" (default), "gml", "kml", "wkt"

Examples

file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
#tib <- tibble::tibble(wkb = vapour_read_geometry(mvfile)) %>%
#  bind_cols(read_gdal_table(mvfile))
pfile <- system.file("extdata/point.shp", package = "vapour")
vapour_read_geometry(pfile)

#> [[1]]
#>  [1] 01 01 00 00 00 00 00 c0 9e b2 f2 e3 3f 00 00 40 b8 86 53 d8 3f
#> 
#> [[2]]
#>  [1] 01 01 00 00 00 00 00 00 47 1c a8 c2 3f 00 00 00 b5 5d 4f d9 3f
#> 
#> [[3]]
#>  [1] 01 01 00 00 00 00 00 80 32 d1 d5 c7 3f 00 00 c0 6a 98 7a d8 3f
#> 
#> [[4]]
#>  [1] 01 01 00 00 00 00 00 00 cb cf 73 e7 3f 00 00 c0 1b 21 34 d8 3f
#> 
#> [[5]]
#>  [1] 01 01 00 00 00 00 00 60 08 18 ad ec 3f 00 00 e0 9a ec 77 e2 3f
#> 
#> [[6]]
#>  [1] 01 01 00 00 00 00 00 c0 40 3c bb d0 3f 00 00 80 0e 30 25 d5 3f
#> 
#> [[7]]
#>  [1] 01 01 00 00 00 00 00 80 3c 35 0d e0 3f 00 00 00 f0 6e 81 a7 3f
#> 
#> [[8]]
#>  [1] 01 01 00 00 00 00 00 40 76 8d a1 ef 3f 00 00 80 70 95 cb e0 3f
#> 
#> [[9]]
#>  [1] 01 01 00 00 00 00 00 80 6a 03 10 d2 3f 00 00 00 58 6a b2 c8 3f
#> 
#> [[10]]
#>  [1] 01 01 00 00 00 00 00 00 d0 05 f3 a9 3f 00 00 40 35 3c f2 df 3f
#>