GDAL raster subdatasets (variables)

A subdataset is a collection abstraction for a number of variables within a single GDAL source. If there's only one variable the datasource and the variable have the same data source string. If there is more than one the subdatasets have the form DRIVER:"datasourcename":varname. Each subdataset name can stand in place of a data source name that has only one variable, so we always treat a source as a subdataset, even if there's only one.

vapour_sds_names(x)

Arguments

x
a data source string, filename, database connection string, Thredds or other URL

Value

list of character vectors, see Details

Details

Returns a list of datasource and subdataset. In the case of a normal data source, with no subdatasets the value of both entries is the datasource.

Examples

f <- system.file("extdata/gdal", "sds.nc", package = "vapour")
## protect from error with netcdf problems
result <- try(vapour_sds_names(f), silent = TRUE)
if (!inherits(result, "try-error")) {
  print(result)
}

#> $datasource
#> [1] "/Users/runner/work/_temp/Library/vapour/extdata/gdal/sds.nc"
#> [2] "/Users/runner/work/_temp/Library/vapour/extdata/gdal/sds.nc"
#> 
#> $subdataset
#> [1] "NETCDF:\"/Users/runner/work/_temp/Library/vapour/extdata/gdal/sds.nc\":vv" 
#> [2] "NETCDF:\"/Users/runner/work/_temp/Library/vapour/extdata/gdal/sds.nc\":vv2"
#> 

vapour_sds_names(system.file("extdata", "sst.tif", package = "vapour"))

#> $datasource
#> [1] "/Users/runner/work/_temp/Library/vapour/extdata/sst.tif"
#> 
#> $subdataset
#> [1] "/Users/runner/work/_temp/Library/vapour/extdata/sst.tif"
#>