vapour

A lightweight GDAL API package for R.

Details

Provides low-level access to 'GDAL' functionality for R packages. The aim is to minimize the level of interpretation put on the 'GDAL' facilities, to enable direct use of it for a variety of purposes. 'GDAL' is the 'Geospatial Data Abstraction Library' a translator for raster and vector geospatial data formats that presents a single raster abstract data model and single vector abstract data model to the calling application for all supported formats https://gdal.org/.

Lightweight means we access parts of the GDAL API as near as possible to their native usage. GDAL is not a lightweight library, but provide a very nice abstraction over format details for a very large number of different formats.

Functions for raster and vector sources are included.

vapour_all_drivers	list of all available drivers, with type and features

vapour_driver	report short name of driver that will be used for a data source

vapour_gdal_version	report version of GDAL in use

vapour_srs_wkt	produce WKT projection string from various projection string inputs

vapour_vsi_list	report contents of VSI sources

vapour_raster_gcp	return internal ground control points, if present

vapour_raster_info	structural metadata of a source

vapour_read_raster	read data direct from a window of a raster band source

vapour_sds_names	list individual raster sources in a source containing subdatasets

vapour_warp_raster	read data direct from a raster source into a specific window

vapour_driver	report name of the driver used for a given source

vapour_geom_name	report attribute name of geometry

vapour_geom_summary	report simple properties of each feature geometry

vapour_layer_names	list names of vector layers in a data source

vapour_layer_info	list of data source, driver, layer name/s, fields, feature count, projection

vapour_read_extent	read the extent, or bounding box, of geometries in a layer

vapour_read_fields	read attributes of features in a layer, the columnar data associated with each geometry

vapour_read_geometry	read geometry in binary (blob, WKB) form

vapour_read_geometry_ia	read geometry by index, arbitrary

vapour_read_geometry_ij	read geometry by sequential index, i to j

vapour_read_geometry_text	read geometry in text form, various formats

vapour_read_names	read the 'names' of features in a layer, the 'FID'

vapour_read_type	read the GDAL types of attributes

vapour_report_fields	report internal type of each attribute by name

As far as possible vapour aims to minimize the level of interpretation provided for the functions, so that developers can choose how things are implemented. Functions return raw lists or vectors rather than data frames or classed types.

options

The following options can be set to control global behaviour.

Sys.getenv("vapour.sql.dialect")	the current SQL dialect in use

SQL dialect

The SQL dialect can be set to "" (empty string), "OGRSQL", or "SQLITE".

The empty string indicates that the native dialect will be used, see OGRSQL and SQLITE for GDAL, accessed 2022-11-11 and the GDAL_DMD_SUPPORTED_SQL_DIALECTS development documentation (since GDAL 3.6).

Setting "NATIVE" as an alias for "" is quite recent and has not been tested with vapour, similarly no testing has been done with non OGRSQL-native or SQLITE-native drivers yet.