Read feature names

Obtains the internal 'Feature ID (FID)' for a data source.

vapour_read_names(dsource, layer = 0L, sql = "")

Arguments

dsource
data source name (path to file, connection string, URL)

layer
integer of layer to work with, defaults to the first (0)

sql
if not empty this is executed against the data source (layer will be ignored)

Details

This may be virtual (created by GDAL for the SQL interface) and may be 0- or 1- based. Some drivers have actual names, and they are persistent and arbitrary. Please use with caution, this function can return the current FIDs, but there's no guarantee of what it represents for subsequent access.

The sql input is only used for the WHERE clause, and forcibly modifies all input query to SELECT FID - use the vapour_read_attributes with SQL to be more specific.

Examples

file <- "list_locality_postcode_meander_valley.tab"
mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
range(fids <- vapour_read_names(mvfile))

#> [1]  1 58

length(fids)

#> [1] 58